refactor pyproject. add support for tui only
